use usize everywhere and some clean-up

+++ src/lib.rs
@@ -12,7 +12,7 @@ pub enum List {
 }
 
 fn calculate_number_words_needed(
-    number_of_words: Option<u8>,
+    number_of_words: Option<usize>,
     minimum_entropy: Option<usize>,
     list_length: usize,
 ) -> usize {
@@ -29,13 +29,14 @@ fn calculate_number_words_needed(
             80
         }
     };
+    // Do the actual math
     let entropy_per_word_from_this_list = (list_length as f64).log2();
     (minimum_entropy as f64 / entropy_per_word_from_this_list).ceil() as usize
 }
 
 /// Actually generate the passphrase, give a couple neccessary parameters.
 pub fn generate_passphrase(
-    number_of_words: Option<u8>,
+    number_of_words: Option<usize>,
     minimum_entropy: Option<usize>,
     separator: &str,
     title_case: bool,
